requirements.txt
attrs==23.1.0
certifi==2023.7.22
contourpy==1.1.1
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.43.1
h11==0.14.0
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.1
openpyxl==3.1.2
outcome==1.2.0
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
plotly==5.17.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tenacity==8.2.3
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.6
wsproto==1.2.0
